Move around some types
