+++ packages/perseus/src/index.ts
@@ -253,3 +253,4 @@ export type {
     TagsShape,
 } from "./multi-items/shape-types";
 export type {Path} from "./multi-items/trees";
+export type {RendererPromptJSON} from "./prompt-types";
