+++ packages/perseus/src/prompt-types.ts
@@ -1,22 +1,8 @@
+import type {RadioPromptJSON} from "./widgets/radio/prompt-utils";
+
 export enum WidgetType {
     RADIO = "RADIO",
 }
-
-export type BasicOption = {
-    value: string;
-};
-
-export type RadioUserInput = {
-    selectedOptions: boolean[];
-    isNoneOfTheAboveSelected: boolean;
-};
-
-export type RadioPromptJSON = {
-    type: WidgetType;
-    hasNoneOfTheAbove: boolean;
-    options: BasicOption[];
-    userInput: RadioUserInput;
-};
 
 export type WidgetPromptJSON = RadioPromptJSON;
 
@@ -26,7 +12,3 @@ export type RendererPromptJSON = {
         [key: string]: WidgetPromptJSON;
     };
 };
-
-export interface GetPromptJSONInterface {
-    getPromptJSON(): RendererPromptJSON;
-}
